hmm1/requirements.txt
# requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-socketio==5.9.0
eventlet==0.33.3

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Security & Crypto
bcrypt==4.0.1
cryptography==41.0.4
PyJWT==2.8.0

# Data validation & serialization
pydantic==2.4.2
marshmallow==3.20.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
python-dateutil==2.8.2